Users can request erasure (#14)
* Basic erasure request system

* Login user, show alert on erasure request confirmation

* Improve confirmation/status UI

* Allow cancelling a current erasure request

Only the user who made the request can cancel it. They will be notified via email to create a paper record.

* Ensure notifications always sent

If a user requests, cancels, and immediately requests again, the second request confirmation email won't be sent, as the ID will be the same, so NotificationSyncer will think it's the same notification. By logging the timestamp, we avoid this.

* No alert necessary for confirming erasure requests